requirements.txt
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-db-connection-pool==1.2.4
Django==4.2.11
drf-spectacular==0.26.2
gevent==22.10.2
gunicorn==20.1.0
python-decouple==3.8
psycopg2-binary==2.9.6
PyJWT==2.6.0
python-dateutil==2.8.2
celery==5.3.1
redis==5.0.0